Render quote list item on search
